requirements.txt
amqp==5.1.1
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
backcall==0.2.0
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==1.0.0
Django==3.2.9
django-celery-beat==2.3.0
django-cors-headers==4.3.0
django-debug-toolbar==3.6.0
django-timezone-field==6.0.1
executing==2.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
kombu==5.3.2
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.16.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.1
traitlets==5.11.2
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
whitenoise==6.2.0
wrapt==1.15.0
